Add files via upload
I recently completed a comprehensive Power BI portfolio project focused on Global Super Store Analysis, and I'm thrilled to showcase the results of my data-driven journey. 

Data Cleaning and Exploration: I meticulously cleaned, organized, and explored the dataset, ensuring that the insights derived were accurate and reliable.

**Dashboard Highlights**:
1. Total Sales: A comprehensive view of total Sales in Store Market on       all Category and Sub-Category.
2. Total Quantity, Profit, Discount, and Country, shipping Cost: Understanding the Total Sales, Total Profit from all items. Total Discount given to all items, Total Country so on...
3.Total Sales Sold by Year by year: Analysis the Total sales By year to year
4. Total Quantity and Sales Sold By each Category: Examining total quantity and sales sold from Category in time to time.
5. Total profit By Market: Identifying the total Profit By each Market.
6. Sales and Profit by Month: Drilling down into monthly statistics.
7. Profit by Country: Graphical Evaluating the Profit achieve of different Country.
8. Sales by Weekdays and Weekends: Insights into which Days the Maximum sales are achieved
9. Total Item Returned: A comprehensive view of total items Returned in Store Market on all Category and Sub-Category.
10. Total Quantities of items Returned: Understanding the Total Quantities returned by all category
11.Total items Returned by Year by year: Analysis the Total price of items returned by year to year
12. Total Quantity and Sales Sold By each Category: Examining total quantity and sales items are returned from Category in time to time.
13. Total items returned By Market: Identifying the total Quantity and items returned by each Market.
This project challenged me to harness the power of Excel for meaningful data visualization and interpretation. The goal was to not only analyse Returned Items from data but also to facilitate informed decision-making for improvements in the sales Items.

Stay tuned for more in-depth insights and a detailed project showcase in the coming weeks. Your feedback and insights are highly appreciated as I continue to refine and expand my data analysis skills!